backend/requirements.txt
fastapi
uvicorn[standard]
requests
sentence-transformers
scikit-learn
numpy
python-dotenv
